SourceCode/requirements.txt
appdirs==1.4.4
# black==21.6b0
certifi==2021.5.30
chardet==4.0.0
click==8.0.1
coloredlogs==15.0
dataclasses==0.6
geojson==2.5.0
humanfriendly==9.1
idna==2.10
importlib-metadata==4.5.0
mypy-extensions==0.4.3
numpy==1.19.5
opencv-contrib-python
opencv-python
paho-mqtt==1.5.1
pathspec==0.8.1
psutil==5.8.0
psycopg2-binary==2.9.1
pyowm==2.10.0
python-dotenv==0.17.1
regex==2021.4.4
requests==2.25.1
screeninfo==0.6.7
toml==0.10.2
tqdm==4.61.0
typed-ast==1.4.3
typing-extensions==3.10.0.0
urllib3==1.26.5
zipp==3.4.1
